adjust for jamin and fix bugs for L8

+++ SIAC/l8_angle.py
@@ -69,30 +69,42 @@ def do_l8_angle(metafile):
     bands = [l8_file_dir + '/' + header + '_%s.TIF'%i for i in bs]
     toa_bands = bands[:-1]
     qa_band  = bands[-1]
-    ang_file = l8_file_dir + '/' + header + '_ANG.txt'
     bands    = np.arange(1, 8)
-    ang_types = ['BOTH', ] + ['SATELLITE',] * 7
-    band_angTypes = zip(bands, ang_types)
-    par = partial(usgs_angle, ang_file = ang_file)
-    cwd = os.getcwd()
-    os.chdir(l8_file_dir + '/angles')
-#     p = Pool() 
-#     rets = p.map(par, band_angTypes) 
-#     p.close()
-#     p.join()  
-#    rets = list(map(par, band_angTypes))
-    rets = [par(i) for i in band_angTypes]
-    sun_angle  = header  + '_solar_B01.img'
-    nsun_angle = header  + '_SAA_SZA.TIF'
-    gdal.Translate(nsun_angle, sun_angle, creationOptions = ['COMPRESS=DEFLATE', 'TILED=YES'], noData='-32767').FlushCache()
-    rets.append(nsun_angle)
-    os.remove(sun_angle)
-    os.remove(sun_angle + '.hdr')
-    os.chdir(cwd)
+    ang_file = l8_file_dir + '/' + header + '_ANG.txt'
     
-    ang_names = [l8_file_dir + '/angles/' + i for i in rets]
+    sun_angle = glob(l8_file_dir + '/angles/' + header  + '_SAA_SZA.TIF')
+    
+    view_angs = []
+    for band in bands:
+        view_ang_file = l8_file_dir + '/angles/' + header  + '_VAA_VZA_B%02d.TIF'%band
+        view_angs += glob(view_ang_file)
+    
+    ang_names = view_angs + sun_angle
+    if len(ang_names) != 8:
+
+        ang_types = ['BOTH', ] + ['SATELLITE',] * 7
+        band_angTypes = zip(bands, ang_types)
+        par = partial(usgs_angle, ang_file = ang_file)
+        cwd = os.getcwd()
+        os.chdir(l8_file_dir + '/angles')
+    #     p = Pool() 
+    #     rets = p.map(par, band_angTypes) 
+    #     p.close()
+    #     p.join()  
+    #    rets = list(map(par, band_angTypes))
+        rets = [par(i) for i in band_angTypes]
+        sun_angle  = header  + '_solar_B01.img'
+        nsun_angle = header  + '_SAA_SZA.TIF'
+        gdal.Translate(nsun_angle, sun_angle, creationOptions = ['COMPRESS=DEFLATE', 'TILED=YES'], noData='-32767').FlushCache()
+        rets.append(nsun_angle)
+        os.remove(sun_angle)
+        os.remove(sun_angle + '.hdr')
+        os.chdir(cwd)
+        ang_names = [l8_file_dir + '/angles/' + i for i in rets]
+
     view_ang_names = ang_names[:-1]
     sun_ang_name = ang_names[-1]
+
     cloud_mask = gdal.Open(qa_band).ReadAsArray()
     cloud_mask = ~((cloud_mask  >= 2720) & ( cloud_mask <= 2732))
